+++ modules/10-gno/codec.go
@@ -1,0 +1,28 @@
+package gno
+
+import (
+	"github.com/cosmos/ibc-go/v10/modules/core/exported"
+
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+)
+
+// RegisterInterfaces registers the gno concrete client-related
+// implementations and interfaces.
+func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
+	registry.RegisterImplementations(
+		(*exported.ClientState)(nil),
+		&ClientState{},
+	)
+	registry.RegisterImplementations(
+		(*exported.ConsensusState)(nil),
+		&ConsensusState{},
+	)
+	registry.RegisterImplementations(
+		(*exported.ClientMessage)(nil),
+		&Header{},
+	)
+	registry.RegisterImplementations(
+		(*exported.ClientMessage)(nil),
+		&Misbehaviour{},
+	)
+}